put all tests in a single file
